Ajuste no CadastroVendedores
