+++ src/App.tsx
@@ -2,9 +2,9 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { CadastroVendedor } from './pages/CadastroVendedores';
-import { CadastroProduto } from './pages/CadastroProduto';
-import { ListagemProdutos } from './pages/ListagemProdutos';
+import { CadastroVendedor } from './pages/SignUp/Layout/CadastroVendedores';
+import { CadastroProduto } from './pages/SignUp/Layout/CadastroProduto';
+import { ListagemProdutos } from './pages/SignUp/Layout/ListagemProdutos';
 
 function App() {
   return (
