+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { CadastroVendedor } from './pages/CadastroVendedores.tsx';
+import { CadastroVendedor } from './pages/SignUp/Layout/CadastroVendedores.tsx';
 import './index.css'
 import App from './App.tsx'
